requirements.txt
boto3==1.34.108
functions_framework
ffmpeg-python